Squashed commit of the following:
commit 88404c565b4cc3fbad899ca1bccd366b43b199b5
Author: AJ Slater <aj@slater.net>
Date:   Sat Oct 3 14:54:40 2020 -0700

    remove TEST variable. always a bad idea

commit a3bef259dc6c72b0a8bb885d0353d7a8bad1b81f
Author: AJ Slater <aj@slater.net>
Date:   Sat Oct 3 13:58:02 2020 -0700

    comments

commit 265784ce9fdf5fe2429fda3fa98f271c2275667b
Author: AJ Slater <aj@slater.net>
Date:   Sat Oct 3 13:54:39 2020 -0700

    optimize select_related for metadata just a little more

commit a488427a15ab8d0efb4a742308a0696d313492dc
Author: AJ Slater <aj@slater.net>
Date:   Sat Oct 3 13:45:15 2020 -0700

    pop fields before serialization. should be faster

commit 07df6a610a7683cf16551c4e644207d10be96f27
Author: AJ Slater <aj@slater.net>
Date:   Sat Oct 3 13:39:42 2020 -0700

    clean up serializers and choices. get choices from src too.

commit 5d59c5a95dcc6194f76d45c578efdd529a895f90
Author: AJ Slater <aj@slater.net>
Date:   Sat Oct 3 00:19:19 2020 -0700

    fix improper null items on undefined metadata fields

commit 905aa375760b374f3e5c8ccdc29f074cf2e3e754
Author: AJ Slater <aj@slater.net>
Date:   Sat Oct 3 00:18:59 2020 -0700

    fix metadata folder view

commit d6017dd76297f448735198e1a3a1c9f111865a2e
Author: AJ Slater <aj@slater.net>
Date:   Fri Oct 2 23:15:54 2020 -0700

    reading choices and messages from the webpack json works now

commit 87246c285ac93674eff1ba55be651a05e967102b
Author: AJ Slater <aj@slater.net>
Date:   Fri Oct 2 14:14:37 2020 -0700

    move item xformer into api

commit 102b291511cc3e6c34d7efe2f4e7839d5637766c
Author: AJ Slater <aj@slater.net>
Date:   Fri Oct 2 14:04:05 2020 -0700

    fix null items

commit b7212c3e3a47987d1b20ba604aba408555b32f70
Author: AJ Slater <aj@slater.net>
Date:   Fri Oct 2 13:44:22 2020 -0700

    sorted items

commit 5c8cef6cd7eb85c246f3099b35a71f2b87555e51
Author: AJ Slater <aj@slater.net>
Date:   Fri Oct 2 13:27:26 2020 -0700

    componentize browser-settings-menu

commit d3fd0f57ea0d5500389e97e1c071896abac3791a
Author: AJ Slater <aj@slater.net>
Date:   Fri Oct 2 13:17:06 2020 -0700

    componentize browser settings dialog

commit 8510179f3be465df0d7848751e62a699902dd458
Author: AJ Slater <aj@slater.net>
Date:   Fri Oct 2 13:05:52 2020 -0700

    working filter menu. componentize top toolbar filters.
