--- codex/choices/readerChoices.json
@@ -1,7 +1,0 @@
-{
-  "fitTo": [
-    { "text": "Height", "value": "HEIGHT" },
-    { "text": "Width", "value": "WIDTH", "default": true },
-    { "text": "Original Size", "value": "ORIG" }
-  ]
-}
