--- codex/choices/websocketMessages.json
@@ -1,10 +1,0 @@
-{
-  "user": {
-    "LIBRARY_CHANGED": "libraryChanged"
-  },
-  "admin": {
-    "SCAN_LIBRARY": "scanLibrary",
-    "FAILED_IMPORTS": "failedImports",
-    "SCAN_DONE": "scanDone"
-  }
-}
