+++ codex/views/mixins.py
@@ -3,9 +3,9 @@ from copy import copy
 
 from django.contrib.sessions.models import Session
 
-from codex.choices.static import DEFAULTS
 from codex.models import Comic
 from codex.models import UserBookmark
+from codex.serializers.webpack import DEFAULTS
 
 
 class SessionMixin:
