+++ frontend/package.json
@@ -17,9 +17,9 @@
     "normalize.css": "^8.0.1",
     "reconnecting-websocket": "^4.4.0",
     "vue": "^2.6.11",
-    "vue-router": "^3.4.4",
+    "vue-router": "^3.4.5",
     "vue2-filters": "^0.11.0",
-    "vuetify": "^2.3.10",
+    "vuetify": "^2.3.12",
     "vuex": "^3.5.1"
   },
   "devDependencies": {
@@ -34,7 +34,7 @@
     "@vue/test-utils": "^1.1.0",
     "babel-core": "^7.0.0-bridge.0",
     "babel-eslint": "^10.1.0",
-    "eslint": "^7.9.0",
+    "eslint": "^7.10.0",
     "eslint-config-adjunct": "^4.8.7",
     "eslint-config-prettier": "^6.12.0",
     "eslint-plugin-array-func": "^3.1.7",
@@ -47,7 +47,7 @@
     "eslint-plugin-markdown": "^1.0.2",
     "eslint-plugin-no-constructor-bind": "^2.0.3",
     "eslint-plugin-no-secrets": "^0.6.8",
-    "eslint-plugin-no-unsanitized": "^3.1.2",
+    "eslint-plugin-no-unsanitized": "^3.1.4",
     "eslint-plugin-no-use-extend-native": "^0.5.0",
     "eslint-plugin-only-warn": "^1.0.2",
     "eslint-plugin-optimize-regex": "^1.2.0",
@@ -59,10 +59,10 @@
     "eslint-plugin-sonarjs": "^0.5.0",
     "eslint-plugin-switch-case": "^1.1.2",
     "eslint-plugin-unicorn": "^22.0.0",
-    "eslint-plugin-vue": "^7.0.0-beta.4",
+    "eslint-plugin-vue": "^7.0.0",
     "eslint-plugin-vue-scoped-css": "^1.0.0",
     "prettier": "^2.1.2",
-    "prettier-plugin-package": "^1.0.0",
+    "prettier-plugin-package": "^1.1.0",
     "sass": "^1.26.11",
     "sass-loader": "^10.0.1",
     "typeface-roboto": "0.0.75",
