+++ frontend/src/browser.vue
@@ -3,126 +3,13 @@
     <header id="browseHeader">
       <v-toolbar id="filterToolbar" class="toolbar" dense>
         <v-toolbar-items v-if="isOpenToSee" id="filterToolbarItems">
-          <v-select
-            ref="filterSelect"
-            v-model="bookmarkFilter"
-            :items="formChoices.bookmark"
-            class="toolbarSelect filterSelect"
-            dense
-            hide-details="auto"
-            ripple
-            :menu-props="{
-              maxHeight: '90vh',
-              overflowY: false,
-              contentClass: filterMenuClass,
-            }"
-            :prepend-inner-icon="filterInnerIcon"
-            @clickHIDE:clear="clearFilters"
-            @click:prepend-inner="clearFilters"
-          >
-            <template #selection="{ item }">
-              {{ item.text }}
-              <span v-if="isOtherFiltersSelected" id="filterSuffix"> + </span>
-            </template>
-            <template #item="data">
-              <v-slide-x-transition hide-on-leave>
-                <v-list-item-content>
-                  <v-list-item-title>
-                    {{ data.item.text }}
-                  </v-list-item-title>
-                </v-list-item-content>
-              </v-slide-x-transition>
-            </template>
-            <template #append-item>
-              <v-slide-x-transition hide-on-leave>
-                <v-divider v-if="filterMode === 'base'" />
-              </v-slide-x-transition>
-              <FilterSubMenu
-                v-for="filterName of filterNames"
-                :key="filterName"
-                :name="filterName"
-                @sub-menu-click="closeFilterSelect"
-              />
-            </template>
-          </v-select>
-          <v-select
-            v-model="rootGroup"
-            :items="rootGroupChoices"
-            class="toolbarSelect rootGroupSelect"
-            dense
-            hide-details="auto"
-            ripple
-          />
-          <v-select
-            v-model="sortBy"
-            class="toolbarSelect sortBySelect"
-            :items="formChoices.sort"
-            prefix="By "
-            dense
-            hide-details="auto"
-            ripple
-          >
-            <template #item="data">
-              <v-list-item v-bind="data.attrs" v-on="data.on">
-                <v-list-item-content>
-                  <v-list-item-title>
-                    {{ data.item.text }}
-                    <v-icon
-                      v-show="sortBy === data.item.value"
-                      class="sortArrow"
-                      :class="{ upsideDown: !settings.sortReverse }"
-                    >
-                      {{ mdiArrowUp }}
-                    </v-icon>
-                  </v-list-item-title>
-                </v-list-item-content>
-              </v-list-item>
-            </template>
-          </v-select>
+          <BrowserFilterSelect />
+          <BrowserRootGroupSelect />
+          <BrowserSortBySelect />
         </v-toolbar-items>
         <v-spacer />
         <v-toolbar-items id="controls">
-          <v-menu offset-y>
-            <template #activator="{ on }">
-              <v-btn icon ripple title="Settings" v-on="on">
-                <v-icon>{{ mdiDotsVertical }}</v-icon>
-              </v-btn>
-            </template>
-            <v-list-item-group id="settingsMenu" ripple>
-              <v-dialog
-                origin="center-top"
-                transition="slide-y-transition"
-                max-width="20em"
-                overlay-opacity="0.5"
-              >
-                <template #activator="{ on }">
-                  <v-list-item v-if="isOpenToSee" v-on="on">
-                    <v-list-item-content>
-                      <v-list-item-title> Settings </v-list-item-title>
-                    </v-list-item-content>
-                  </v-list-item>
-                </template>
-                <h3>Browser Settings</h3>
-                <div id="settingsGroupCaption" class="text-caption">
-                  Show these groups when navigating the browse heirarchy.
-                </div>
-                <v-checkbox
-                  v-for="choice of formChoices.settingsGroup"
-                  :key="choice.value"
-                  :input-value="getShow(choice.value)"
-                  :label="`Show ${choice.text}`"
-                  dense
-                  @change="setShow(choice.value, $event)"
-                />
-              </v-dialog>
-              <AuthDialog />
-              <v-list-item v-if="isOpenToSee" @click="reload">
-                <v-list-item-content>
-                  <v-list-item-title> Reload Libraries</v-list-item-title>
-                </v-list-item-content>
-              </v-list-item>
-            </v-list-item-group>
-          </v-menu>
+          <BrowserSettingsMenu />
         </v-toolbar-items>
       </v-toolbar>
       <v-toolbar id="titleToolbar" class="toolbar" dense>
@@ -151,7 +38,7 @@
       </div>
     </v-main>
     <v-main v-else-if="itemsExist" id="browsePane">
-      <BrowseCard
+      <BrowserCard
         v-for="item in objList"
         :key="`${item.group}${item.pk}`"
         :item="item"
@@ -242,16 +129,18 @@
 </template>
 
 <script>
-import { mdiArrowUp, mdiCloseCircle, mdiDotsVertical } from "@mdi/js";
+import { mdiArrowUp, mdiDotsVertical } from "@mdi/js";
 import { mapGetters, mapState } from "vuex";
 
 import { ADMIN_URL } from "@/api/v1/auth";
 import { FAILED_IMPORT_URL, getSocket } from "@/api/v1/browser";
 import WS_MESSAGES from "@/choices/websocketMessages.json";
-import AuthDialog from "@/components/auth-dialog";
-import BrowseCard from "@/components/browse-card";
+import BrowserCard from "@/components/browser-card";
+import BrowserFilterSelect from "@/components/browser-filter-select";
+import BrowserRootGroupSelect from "@/components/browser-root-group-select";
+import BrowserSettingsMenu from "@/components/browser-settings-menu";
+import BrowserSortBySelect from "@/components/browser-sort-by-select";
 import { getVolumeName } from "@/components/comic-name";
-import FilterSubMenu from "@/components/filter-sub-menu";
 import PlaceholderLoading from "@/components/placeholder-loading";
 
 const SCAN_DONE_MSGS = new Set([
@@ -262,9 +151,11 @@ const SCAN_DONE_MSGS = new Set([
 export default {
   name: "Browser",
   components: {
-    AuthDialog,
-    BrowseCard,
-    FilterSubMenu,
+    BrowserCard,
+    BrowserFilterSelect,
+    BrowserRootGroupSelect,
+    BrowserSettingsMenu,
+    BrowserSortBySelect,
     PlaceholderLoading,
   },
   data() {
@@ -281,11 +172,8 @@ export default {
     ...mapState("browser", {
       browseTitle: (state) => state.browseTitle,
       currentRoute: (state) => state.routes.current,
-      formChoices: (state) => state.formChoices,
-      settings: (state) => state.settings,
       upRoute: (state) => state.routes.up,
       objList: (state) => state.objList,
-      filterMode: (state) => state.filterMode,
       browseLoaded: (state) => state.browseLoaded,
       librariesExist: (state) => state.librariesExist,
       itemsExist: (state) => state.objList && state.objList.length > 0,
@@ -298,7 +186,6 @@ export default {
       enableNonUsers: (state) => state.enableNonUsers,
     }),
     ...mapGetters("auth", ["isAdmin", "isOpenToSee"]),
-    ...mapGetters("browser", ["rootGroupChoices", "filterNames"]),
     showPlaceHolder: function () {
       return (
         this.enableNonUsers === undefined ||
@@ -353,89 +240,8 @@ export default {
       }
       return "";
     },
-    filterInnerIcon: function () {
-      if (this.isFiltersClearable) {
-        return mdiCloseCircle;
-      }
-      return " ";
-    },
-    filterMenuClass: function () {
-      // Lets me hide bookmark menu items with css when the filterMode
-      //   changes.
-      let className = "filterMenu";
-      if (this.filterMode !== "base") {
-        className += "Hidden";
-      }
-      return className;
-    },
-    isOtherFiltersSelected: function () {
-      for (let filterName of this.filterNames) {
-        const filterArray = this.settings.filters[filterName];
-        if (filterArray && filterArray.length > 0) {
-          return true;
-        }
-      }
-      return false;
-    },
-    isFiltersClearable: function () {
-      if (this.bookmarkFilter !== "ALL") {
-        return true;
-      }
-      return this.isOtherFiltersSelected;
-    },
     showUpButton: function () {
       return this.upRoute && this.upRoute.group;
-    },
-    bookmarkFilter: {
-      get() {
-        return this.getFilter("bookmark");
-      },
-      set(value) {
-        if (value === null || value === undefined) {
-          console.warn(`bookmarkFilter was ${value}. Setting to 'ALL'`);
-          value = "ALL";
-        }
-        this.setFilter("bookmark", value);
-      },
-    },
-    decadeFilter: {
-      get() {
-        return this.getFilter("decade");
-      },
-      set(value) {
-        this.setFilter("decade", value);
-      },
-    },
-    charactersFilter: {
-      get() {
-        return this.getFilter("characters");
-      },
-      set(value) {
-        this.setFilter("characters", value);
-      },
-    },
-    rootGroup: {
-      get() {
-        return this.settings.rootGroup;
-      },
-      set(value) {
-        this.settingChanged({ rootGroup: value });
-      },
-    },
-    sortBy: {
-      get() {
-        return this.settings.sortBy;
-      },
-      set(value) {
-        let data = {};
-        if (value === this.settings.sortBy) {
-          data.sortReverse = !this.settings.sortReverse;
-        } else {
-          data.sortReverse = false;
-          data.sortBy = value;
-        }
-        this.settingChanged(data);
-      },
     },
   },
   watch: {
@@ -484,37 +290,6 @@ export default {
         }
       }
     },
-    settingChanged: function (data) {
-      this.$store.dispatch("browser/settingChanged", data);
-    },
-    setSubSetting: function (key, subKey, value) {
-      const data = { [key]: { [subKey]: value } };
-      this.settingChanged(data);
-    },
-    getShow: function (group) {
-      return this.settings.show[group];
-    },
-    setShow: function (group, value) {
-      this.setSubSetting("show", group, value === true);
-    },
-    getFilter: function (filterName) {
-      return this.settings.filters[filterName];
-    },
-    setFilter: function (filterName, value) {
-      this.setSubSetting("filters", filterName, value);
-    },
-    clearFilters: function () {
-      this.$store.dispatch("browser/clearFilters");
-    },
-    closeFilterSelect: function () {
-      // On click, reselect the current value to close the menu.
-      const item = this.$refs.filterSelect.selectedItems[0];
-      this.$refs.filterSelect.selectItem(item);
-      this.$store.dispatch("browser/setFilterMode", { mode: "base" });
-    },
-    reload: function () {
-      // this.$store.dispatch("browser/browseOpened", this.$route.params);
-    },
     routeToPage: function (page) {
       const route = {
         name: this.$route.name,
@@ -553,29 +328,14 @@ export default {
 #titleToolbar .v-toolbar__title {
   margin: auto;
 }
-.toolbarSelect {
-}
 #browsePane {
   display: flex;
   margin-top: 96px;
   margin-left: 15px;
   overflow: auto;
 }
-.sortArrow {
-  width: 20px;
-  float: right;
-}
-.upsideDown {
-  transform: rotate(180deg);
-}
 .invisible {
   visibility: hidden;
-}
-#filterSuffix {
-  margin-left: 0.25em;
-}
-#settingsGroupCaption {
-  color: gray;
 }
 #noComicsFound {
   font-size: x-large;
@@ -594,20 +354,11 @@ export default {
 #dismissScanNotify {
   margin-left: 5px;
 }
-#settingsMenu {
-  background-color: #121212;
-}
-#progressBar {
-  margin-top: 40px;
-}
 #announce {
   text-align: center;
 }
 .outdated {
   font-style: italic;
-}
-
-#failedImportsNotify {
 }
 #failedImportsNotify a {
   padding-right: 0.5em;
@@ -654,14 +405,6 @@ export default {
 .filterMenuHidden > .v-select-list > .v-list-item {
   display: none !important;
 }
-.rootGroupSelect {
-  width: 118px;
-  margin-left: 8px;
-  margin-right: 8px;
-}
-.sortBySelect {
-  width: 168px;
-}
 #scanNotify > .v-snack__wrapper {
   min-width: 183px;
 }
@@ -676,15 +419,6 @@ export default {
   }
   .filterSelect .v-input__icon--prepend-inner svg {
     width: 13px;
-  }
-  .rootGroupSelect {
-    width: 95px;
-    margin-left: 4px;
-    margin-right: 4px;
-  }
-  .sortBySelect {
-    width: 133px;
-    margin-right: -19px;
   }
   #titleToolbar .v-toolbar__title {
     font-size: 0.75rem;
