+++ frontend/vue.config.js
@@ -18,6 +18,12 @@ module.exports = {
         },
       }),
     ],
+    entry: {
+      //app: './src/main.js',
+      browserChoices: "./src/choices/browserChoices.json",
+      readerChoices: "./src/choices/readerChoices.json",
+      websocketMessages: "./src/choices/websocketMessages.json",
+    },
     devServer: {
       writeToDisk: true, // Write files to disk in dev mode, so Django can serve the assets
     },
