+++ setup-dev.sh
@@ -9,6 +9,3 @@ poetry install --no-root
 
 echo "*** install frontend build dependencies ***"
 bash -c "cd frontend && npm install"
-
-echo "*** copy choices from frontend ***"
-cp frontend/src/choices/* codex/choices/
